+++ lib/src/model/page_view_model.dart
@@ -3,27 +3,33 @@ import 'package:introduction_screen/src/model/page_decoration.dart';
 
 class PageViewModel {
   /// Title of page
-  final String title;
+  final String? title;
 
   /// Title of page
-  final Widget titleWidget;
+  final Widget? titleWidget;
 
   /// Text of page (description)
-  final String body;
+  final String? body;
 
   /// Widget content of page (description)
-  final Widget bodyWidget;
+  final Widget? bodyWidget;
 
   /// Image of page
   /// Tips: Wrap your image with an alignment widget like Align or Center
-  final Widget image;
+  final Widget? image;
 
   /// Footer widget, you can add a button for example
-  final Widget footer;
+  final Widget? footer;
 
   /// Page decoration
   /// Contain all page customizations, like page color, text styles
   final PageDecoration decoration;
+
+  /// If widget Order is reverse - body before image
+  final bool reverse;
+
+  /// Wrap content in scrollView
+  final bool useScrollView;
 
   PageViewModel({
     this.title,
@@ -32,7 +38,9 @@ class PageViewModel {
     this.bodyWidget,
     this.image,
     this.footer,
+    this.reverse =  false,
     this.decoration = const PageDecoration(),
+    this.useScrollView = true,
   })  : assert(
           title != null || titleWidget != null,
           "You must provide either title (String) or titleWidget (Widget).",