+++ lib/src/ui/intro_button.dart
@@ -1,20 +1,24 @@
 import 'package:flutter/material.dart';
 
 class IntroButton extends StatelessWidget {
-  final VoidCallback onPressed;
-  final Widget child;
-  final Color color;
+  final VoidCallback? onPressed;
+  final Widget? child;
+  final Color? color;
 
-  const IntroButton({Key key, this.onPressed, @required this.child, this.color})
+  const IntroButton({Key? key, this.onPressed, this.child, this.color})
       : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    return FlatButton(
+    return TextButton(
       onPressed: onPressed,
-      child: child,
-      color: color,
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(8.0)),
+      child: child ?? SizedBox(),
+      style: TextButton.styleFrom(
+        primary: color,
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(8.0),
+        ),
+      ),
     );
   }
 }